Refactored tests to create a MockTestRunner that can help with checking the results of test runner modifications.
